pyopengl2-92 refactor GLFW/GL context specification result
implement __lock inside `OGLMetaEntity`
    `get_concrete` has to be locked to ensure only needed entities are created.

    As using 'unique render thread' approach, its users obligation to control timing correctly. But as rendering often has sleeping phase, think threading can bring some benefit in terms of efficiency.
    resolve #92
