+++ src/wkernel/devices/render/frame/base.py
@@ -170,7 +170,7 @@ class FrameRenderer:
         self.__pane_vrtx_block['tex_coord'] = (xs, ys), (xe, ys), (xe, ye), (xs, ye)
         self.__pane_vbo.push_cache()
         # render
-        with texture:
+        with texture as t:
             with self.__pane_vao as vao:
                 with self.__pane_prgrm:
                     gl.glDrawArrays(gl.GL_QUADS, 0, 4)
